requirements.txt
autopep8==2.0.1
blinker==1.5
click==8.1.3
Flask==2.2.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-WhooshAlchemy==0.56
Flask-WTF==1.1.1
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.33
protobuf==3.20.3
pycodestyle==2.10.0
PyMySQL==1.0.2
searchform==0.1.1
SQLAlchemy==1.4.46
tomli==2.0.1
Werkzeug==2.2.2
Whoosh==2.7.4
WTForms==3.0.1
zipp==3.11.0
